Added saving (& sharing) of compositions.
Again needs styling, and could I'm not sure that the fading text-box is such a good way of doing it.

The method used to create a psuedo-random field but allow saving and storing, without storing the whole grid, is described in the Node constructor.

Also made the major scale the default - it's much happier - and increased the size of the scales for more available notes.